fix: use Long for note ID throughout persistence layer

+++ androidApp/src/main/kotlin/com/jfranco/sharetosave/persistence/entity/NoteEntity.kt
@@ -8,7 +8,7 @@ import java.time.LocalDateTime
 
 @Entity(tableName = "note")
 data class NoteEntity(
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id") val id: Int,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id") val id: Long,
     @ColumnInfo(name = "title") val title: String?,
     @ColumnInfo(name = "content") val content: String?,
     @ColumnInfo(name = "image") val image: String?,
